location innit

+++ game_data.py
@@ -144,7 +144,7 @@ class Location:
     first_visit: bool
     available_items: Optional[list]
 
-    def __init__(self, name : str, num : int, points : int, long : str, short : str, items : Optional[list] = None) -> None:
+    def __init__(self, name : str, num : int, points : int, short : str, long : str, items : Optional[list] = None) -> None:
         """Initialize a new location.
 
         # TODO Add more details here about the initialization if needed
